+++ src/App.vue
@@ -1,5 +1,5 @@
 <script setup>
-import NavBar from './assets/components/NavBar.vue'
+import NavBar from '@/components/NavBar.vue'
 </script>
 
 <template>
