+++ src/components/Carousel.vue
